refactor: consolidate enums
