+++ action.py
@@ -1,11 +1,5 @@
 from manifest import ManifestItem
-from enum import Enum
 
-class ActionTypes(Enum):
-    FromPark = 1
-    ToItem = 2
-    MoveItem = 3
-    ToPark = 4
 
 class Action:
     def __init__(self, source: ManifestItem, target: ManifestItem):
