+++ grid_display.py
@@ -1,8 +1,11 @@
-from cell import Cell, ParkCell, CellTypes, TargetTypes
+from PySide6 import QtWidgets
+
+from action import Action
+from cell import Cell, ParkCell
+from enums import ActionTypes, TargetTypes
 from manifest import ManifestItem
 from state import State
-from action import Action, ActionTypes
-from PySide6 import QtWidgets
+
 
 class GridDisplay(State):
     def __init__(self, grid: list[list[ManifestItem]], parkLabel:QtWidgets.QLabel):
@@ -33,7 +36,7 @@ class GridDisplay(State):
 
                 cell.set_item(item)
                 cell._set_label_text()
-                cell.set_type(CellTypes.to_type(item.get_title()))
+                cell.set_type(item.get_type())
                 cell.set_targetType(None)
         
         if source.coordinate == self.parkCell.coordinate:
